Allow altering the command executed
This allows us to override the test executed by buster-autotest using a "cmd" property in the options hash. This allows reusing the watching / re-execute logic when you have a custom version of buster-test. 

e.g. in my case I have an executable called "buster-options-test" (allowing v8 options) and a matching "buster-options-autotest" which passes in the appropriate command { cmd: "buster-options-test" }